# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shake Crazy Food Truck API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    get:
      operationId: getApprovedSubmissions
      tags: [submissions]
      summary: Get approved user submissions
      responses:
        "200":
          description: Approved submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit a love sharing entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionRequest"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
  /admin/submissions:
    get:
      operationId: adminGetSubmissions
      tags: [admin]
      summary: Get all submissions (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
  /admin/submissions/{id}/approve:
    put:
      operationId: adminApproveSubmission
      tags: [admin]
      summary: Approve/reject submission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveSubmissionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Submission:
      type: object
      properties:
        id:
          type: integer
        userName:
          type: string
        imageUrl:
          type: string
          nullable: true
        experienceText:
          type: string
        approved:
          type: boolean
        location:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userName
        - experienceText
        - approved
        - createdAt
    SubmissionRequest:
      type: object
      properties:
        userName:
          type: string
        imageUrl:
          type: string
          nullable: true
        experienceText:
          type: string
        location:
          type: string
          nullable: true
      required:
        - userName
        - experienceText
    ApproveSubmissionRequest:
      type: object
      properties:
        approved:
          type: boolean
      required:
        - approved
...