# lib/api-zod/src/generated/types/approveSubmissionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shake Crazy Food Truck API
 * OpenAPI spec version: 0.1.0
 */

export interface ApproveSubmissionRequest {
  approved: boolean;
}