# lib/api-zod/src/generated/types/submissionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shake Crazy Food Truck API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmissionRequest {
  userName: string;
  imageUrl?: string | null;
  experienceText: string;
  location?: string | null;
}